requirements.txt
pygal
Pillow
pygame
easygui
cairosvg
